small fix for bti data related to r3186: convert the (sparse) data matrix back to double
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@3194 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ fileio/ft_read_data.m
@@ -380,7 +380,7 @@ switch dataformat
         error('unsupported data format');
     end
     % calibrate the data
-    dat = sparse(calib)*dat;
+    dat = double(sparse(calib)*dat);
 
   case 'bci2000_dat'
     % this requires the load_bcidat mex file to be present on the path
